Custom Bots/Data Entry Bot/requirements.txt
pandas==2.0.0
openpyxl==3.1.2
selenium==4.10.0
webdriver-manager==3.8.6
requests==2.31.0
python-dotenv==1.0.0 